Move low-level functions from TilemapRenderer into a base class
