+++ bsnes/ui-qt/ui-base.hpp
@@ -48,9 +48,11 @@ using namespace ruby;
   #include "debugger/tools/memory.moc.hpp"
   #include "debugger/tools/properties.moc.hpp"
 
+  #include "debugger/ppu/base-renderer.hpp"
+  #include "debugger/ppu/tilemap-renderer.hpp"
+
   #include "debugger/ppu/cgram-widget.moc.hpp"
   #include "debugger/ppu/image-grid-widget.moc.hpp"
-  #include "debugger/ppu/tilemap-renderer.hpp"
 
   #include "debugger/ppu/vram-viewer.moc.hpp"
   #include "debugger/ppu/tilemap-viewer.moc.hpp"
